// x-pack/platform/plugins/shared/fleet/public/applications/integrations/sections/epm/screens/home/card_utils.d.ts
import React from 'react';
import type { CustomIntegration, CustomIntegrationIcon } from '@kbn/custom-integrations-plugin/common';
import type { DeprecationInfo, EpmPackageInstallStatus, IntegrationCardReleaseLabel, PackageSpecIcon } from '../../../../../../../common/types';
import type { DynamicPage, DynamicPagePathValues, StaticPage } from '../../../../constants';
import type { PackageListItem } from '../../../../types';
export interface IntegrationCardItem {
    categories: string[];
    description: string;
    descriptionLineClamp?: number;
    extraLabelsBadges?: React.ReactNode[];
    fromIntegrations?: string;
    hasDataStreams?: boolean;
    icons: Array<PackageSpecIcon | CustomIntegrationIcon>;
    id: string;
    installStatus?: EpmPackageInstallStatus;
    integration: string;
    isCollectionCard?: boolean;
    isQuickstart?: boolean;
    isReauthorizationRequired?: boolean;
    isUnverified?: boolean;
    isUpdateAvailable?: boolean;
    isDeprecated?: boolean;
    deprecationInfo?: DeprecationInfo;
    maxCardHeight?: number;
    minCardHeight?: number;
    name: string;
    onCardClick?: () => void;
    release?: IntegrationCardReleaseLabel;
    showDescription?: boolean;
    showInstallationStatus?: boolean;
    showCompressedInstallationStatus?: boolean;
    showLabels?: boolean;
    showReleaseBadge?: boolean;
    title: string;
    titleLineClamp?: number;
    titleBadge?: React.ReactNode;
    titleSize?: 'xs' | 's';
    url: string;
    version: string;
    type?: string;
    supportsAgentless?: boolean;
    signalTypes?: string[];
}
export declare const mapToCard: ({ getAbsolutePath, getHref, item, addBasePath, packageVerificationKeyId, selectedCategory, }: {
    getAbsolutePath: (p: string) => string;
    getHref: (page: StaticPage | DynamicPage, values?: DynamicPagePathValues) => string;
    addBasePath: (url: string) => string;
    item: CustomIntegration | PackageListItem;
    packageVerificationKeyId?: string;
    selectedCategory?: string;
}) => IntegrationCardItem;
export declare function getIntegrationLabels(item: PackageListItem): React.ReactNode[];